Added a package manager for installing programs
Currently there are no programs besides test2 but yeah do install
program and then the name and it will install it from my server. To
submit a program go to my discord thats in the readme and put it in the
programs channel. I'll manually review it and see if I should put it on
my server.